+++ conf/url_redirects.py
@@ -13,4 +13,9 @@ redirects = [
         QuerystringRedirectView.as_view(pattern_name='brexit-international-contact-form-success'),
         name='eu-exit-international-contact-form-success'
     ),
+    url(
+        r'^international/content/industries/advanced-manufacturing/$',
+        QuerystringRedirectView.as_view(url='/international/content/industries/engineering-and-manufacturing/'),
+        name='advanced-manufacturing-redirect'
+    ),
 ]